+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -7,11 +7,12 @@
 - Added support of features that are known into the future/at forecast time for dnn in AutoML Forecasting jobs.
 - Added support for new workspace connection types: azure_open_ai, cognitive_search, and cognitive_service.
 - Added support for new credential type: ApiKeyConfiguration.
+- Added support of `download` for component operations.
 
 ### Bugs Fixed
 - Local job runs will no longer fail if Docker registry has no username/password
-
 - Fixed an issue that code asset doesn't work with relative symbol links.
+- Fixed [Issue 31319](https://github.com/Azure/azure-sdk-for-python/issues/31319): can't accept `PathLike` for `CommandComponent.code`.
 
 ### Breaking Changes
 
@@ -21,7 +22,7 @@
   (instead of the default behavior for python < 3.15, which uses locale specific encodings)
 - Removed references to deprecated "feature_store" workspace connection type.
 
-## 1.9.0 (Unreleased)
+## 1.9.0 (2023-07-25)
 
 ### Features Added
 - Added support to enable gpu access (local_enable_gpu) for local deployment.